+++ crates/services/relayer/tests/integration.rs
@@ -19,6 +19,7 @@ use fuel_core_relayer::{
         LogTestHelper,
     },
     Config,
+    H160,
 };
 use fuel_core_services::Service;
 
@@ -284,7 +285,8 @@ impl TestContext {
     }
 
     fn given_logs(&mut self, mut logs: Vec<Log>) {
-        let contract_address = self.config.eth_v2_listening_contracts[0];
+        let contract_address =
+            H160::from_slice(self.config.eth_v2_listening_contracts[0].as_slice());
         for log in &mut logs {
             log.address = contract_address;
         }